ajout de nouveau manga + changement de type pour les note dans la table MangaLectureFini

+++ MyMangaList/database/migrations/2021_06_14_132334_create_MangaLectureEnCours_table.php
@@ -16,7 +16,7 @@ class CreateMangaLectureEnCoursTable extends Migration
         Schema::create('MangaLectureEnCours', function (Blueprint $table) {
             $table->unsignedBigInteger('Id_Utilisateur');
             $table->unsignedBigInteger('Id_Manga')->index('Id_Manga');
-            $table->integer('note')->nullable();
+            $table->float('note')->nullable();
             $table->integer('tomeEnCours')->nullable();
             $table->integer('chapitreEnCours')->nullable();
             $table->primary(['Id_Utilisateur', 'Id_Manga']);
